+++ packages/framework/src/Form/Transformers/ImagesIdsToImagesTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Transformers;
+namespace Shopsys\FrameworkBundle\Form\Transformers;
 
 use IteratorAggregate;
 use Shopsys\FrameworkBundle\Component\Image\ImageFacade;
