+++ packages/framework/src/Form/Transformers/ProductsIdsToProductsTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Transformers;
+namespace Shopsys\FrameworkBundle\Form\Transformers;
 
 use IteratorAggregate;
 use Shopsys\FrameworkBundle\Model\Product\ProductRepository;
