+++ packages/framework/src/Form/Transformers/ScriptPlacementToBooleanTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Transformers;
+namespace Shopsys\FrameworkBundle\Form\Transformers;
 
 use Shopsys\FrameworkBundle\Model\Script\Script;
 use Symfony\Component\Form\DataTransformerInterface;
